+++ src/pages/home/home.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Home extends Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div></div>
+        );
+    }
+}
